Added fake compress method
This 'compression' method does not actually compressed the source file at all,
but rather inserts the required headers into the stream of data so that the
decompression method can return the original file, which still achieves the
desired result. As a result, the 'compressed' data is actually larger than the
uncompressed.

For more information, see:
<https://reverseengineering.stackexchange.com/questions/16021/what-compression-algorithm-is-this>